Move assert corruption logic into TestParams struct for improved code organization
